requirements.txt
websocket-client==1.2.1
asyncio==3.4.3